requirements.txt
# Requirements for desktop-only version
Flask>=2.0
pathspec>=0.9
detect-secrets>=1.4.0
requests>=2.20
pyperclip
Flask-SocketIO
pywebview
appdirs
python-docx>=0.8.11
reportlab>=3.6.0
markdown>=3.4.0